Move NetworkSession to separate file.
